+++ src/objectdetection/utils/tensor-node-utils.ts
@@ -41,6 +41,7 @@ export class TensorNodeUtils {
             }
 
             const tf = require('@tensorflow/tfjs');
+            tf.disableDeprecationWarnings();
             const fileSystem = require('@tensorflow/tfjs-node/dist/io/file_system');
             if (tf.io.getLoadHandlers('file://').length === 0) {
                 tf.io.registerLoadRouter(fileSystem.nodeFileSystemRouter);